requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
requests==2.32.3
PyJWT[crypto]==2.9.0
google-generativeai
azure-storage-blob
azure-data-tables
pydantic==2.9.2
pyyaml==6.0.2
azure-identity==1.18.0
